work on wt drum brain

+++ lib/protar_neck/protar_neck_device.hpp
@@ -44,8 +44,8 @@ private:
     SPIMasterInterface interface;
     uint8_t m_attPin;
     bool valid = false;
-    long last = 0;
-    long lastInput = 0;
-    long lastInit = 0;
+    uint32_t last = 0;
+    uint32_t lastInput = 0;
+    uint32_t lastInit = 0;
     protarneck_t lastInputs;
 };
